Reduce rate limiting
The original rate was set by increasing it until the functional tests
running on preview no longer triggered any 429s.

We forgot that in production we actually run three instances of the
router. This means that effectively the rate limit it triple what we
originally set (preview only runs one router instance).

We scaled up staging to the same number and size of instances as preview
and ran some load tests against it.

Using `hey` (https://github.com/rakyll/hey) from the Jenkins box to
allow for maximum network connectivity (Jenkins is in the same AWS
region as PaaS infrastructure), we could run tests with a rate limit of
up to 146r/s without rate limiting kicking in.

The other issue that we've seen is that under heavy load, PaaS shut down
our instances. This was due to the health check they send timing out.
This results in the instance being deemed unhealthy and so shut down.
The healthchecks are sent every 30 seconds, and the response must take
less than 1 second.

We've seen instances failing this healthcheck with load at 111r/s. When
the instance is shut down, any requests in process are returned as 502s.

We need to set the rate limiting at a level that we start returning 429s
well before the level of traffic that can cause these 502s.