proxy_pass: make sure, in most cases, target url is specified with a trailing slash
this ensures that the target is interpreted as a URI and so has the *normalized*
path of the incoming request applied to it to generate the onwards path, notably
having performed slash merging.

this does however mean that we can't use nginx "variables" for these target urls
because those are too dynamic and opaque to nginx's configuration system for
it to be able to make decisions about behaviour based on their values.
fortunately we can of course still use jinja variables.

also as a result this means the test variables need to refer to real existing
hosts because nginx is now able to check these hosts at init time, so it will.