592-504-errors-on-production-nginx-dns
We need to define our urls as variables and pass these to proxy pass to ensure the dns lookup is periodically expired.
We have seen 504 errors when the ips of our s3 buckets are changed because the DNS resolution is never expired and always points to the ip looked up when the papp is started.

> domain names statically configured in config are only looked up once on startup (or configuration reload).
https://forum.nginx.org/read.php?2,215830,215832#msg-215832

> Setting proxy_pass to a variable forces re-resolution of the DNS names, because Nginx treats variables differently to static configuration
https://serverfault.com/a/593003

These variables must be set _outside_ the location block:
https://serverfault.com/questions/240476/how-to-force-nginx-to-resolve-dns-of-a-dynamic-hostname-everytime-when-doing-p#comment875954_593003