supervisord: attempt more graceful shutdown procedure through better use of signals
we will never make this perfect as it's possible for an instance to have
a significant number of concurrent or queued connections and only a 10s
grace period given by bosh in which to clear that backlog. so make sure
daemons are given graceful shutdown signals so they can get on with doing
that backlog-clearing. ensure awslogs daemon doesn't shut down at all until
the last second so we have a good shot at getting logs of all handled
requests.