disable gzip at toplevel
there isn't a practical way of disabling gzip for text/html in nginx, and
the sensitive nature of some of the html means we probably don't want
compression there due to e.g. BREACH concerns. instead we will perform
gzipping in the applications where we can have a little more control.