Use nginx.retries to ask nginx-prometheus-exporter to wait for NGINX
We've had issues where the router app metrics have not been available
because the nginx-prometheus-exporter app has started before the NGINX
stub_status page is available, and exited with an error.

This commit uses the command line argument `-nginx.retries int` to get
nginx-prometheus-exporter to retry a couple of times before exiting with
an error.

3 seems like as good a number of retries as any.