Add supervisord process to forward NGINX logs to stdout
We want to be able to see NGINX access logs in stdout of our PaaS app,
so it will appear in `cf logs` and be forwarded to Splunk.

This commit copies the method used in
alphagov/digitalmarketplace-docker-base#74, adding a process that
continually echoes the contents of the log file to stdout. We need to
keep the logfile so that awslogs can keep forwarding to CloudWatch.