Upgrade to Python 3.7
The router uses Python to generate nginx config files on startup,
but Python 2.7 was EOL on January 1st 2020 and needs upgrading.

Install Python 3.7 instead of 2.7, update dependencies and add
`python3` to the hashbang lines in the python scripts. There is more
cleanup work that can be done in this app (see [Trello](https://trello.com/c/kQ6vKiRy/1162-3-manage-router-docker-image-in-docker-base-repo-split-me)
for more details), but that's out of scope for this PR.

Note that this Docker image is hard to test. I have run the unit tests
and successfully built and run the iamge locally. I've also logged
into the container and checked that the environment variables
have been injected into the `.conf` files everywhere they're
referenced.

https://trello.com/c/hwA5XVWT/1497-upgrade-the-router-off-python-27