Reduce rate limit to 5 requests/second/instance
We have a scraper that hits us with a very large amount of traffic in a short space of time every day. Buyer-frontend is unable to cope, and gets terminated.

The scraper is currently just under our rate limit. So try reducing the rate limit. Hopefully, they are nice and will start limiting their request rate to one we can handle. If not, we may need to try more extreme measures.

I propose lowering the rate limit to 5 requests/second/instance (30 requests/second overall). We should review our logs after a week to see what impact this had on users.