Allow inline scripts, stylesheets and SVG images
We use inline scripts to initialise the GOV.UK frontend, for some
custom Google Analytics tracking and on our router templates for
maintenance mode. Unfortunately, there are too many scripts to
include all the hashes in the header. This does mean we lose some
of the benefit from the CSP header, but not all.