Reduce request rate limits and allowed burst
We've got a scraper who attempts to scrape 1500-2000 pages in the space of 2 minutes on a regular basis. We're almost able to cope with this, but are still getting the odd 502 response that suggests that the DMp is unable to cope with this level of traffic 100% reliably.

So halve the permitted rate limit and also the allowed burst. The DMp should easily be able to handle the scraper's current level of traffic given this rate limit. We'll need to monitor to check whether this affects any other users.

We also need to take other steps to make the DMp more able to cope with this sort of traffic. Reducing the rate limit should buy us the time to make those improvements.